Removed '.git*' from '/usr/src/app'
Due to the COPY command the local '.git*' files and dirs are also copied.
This can cause problems when running 'npm install', especially when npm needs to use git to fetch packages and the `graphql-server` is a git submodule. As only the submodule is copied into the Docker image, but not the containing parent git project, git is confused and exits with an error.