Using a different dir in Docker
graphql-server is now mounted into /usr/ScienceDbStarterPack/graphql-server
This is to enable make npm install work in the context of ScienceDbStarterPack.
The problem occurred when graphql-server is managed as a git submodule and npm install
invokes git to fetch some remote npm packages. Git exits with an error, if the parent project
including graphql-server as a submodule is not present. Tried many solutions. The working
one is to mount actually the whole project and thus satisfy git's requirements.